Enable scheduler to be run from installed package
The scheduler uses some files packaged with the project
distribution and thus need to consider all possible ways
to start the service to retrieve these files. Using the
python pkg_resources module, this is taken care and works
both from a development environment as well as a installed
package scenario.